+++ src/routes/currency/[id]/+page.svelte
@@ -18,10 +18,12 @@
 	import Note from '$lib/components/Note.svelte';
 	import Tokens from '$lib/components/Tokens.svelte';
 	import Link from '$lib/components/Link.svelte';
+	import Slider from '$lib/components/Slider.svelte';
+	import Checkbox from '$lib/components/Checkbox.svelte';
 
 	let { data }: { data: PageData } = $props();
 
-	let { currency, venues, count } = $derived(data);
+	let { currency, venues, count, scholarCount, venueCount } = $derived(data);
 
 	const db = getDB();
 	const auth = getAuth();
@@ -30,34 +32,26 @@
 	let isMinter = $derived(currency && user && currency.minters.includes(user));
 
 	let newMinter = $state('');
-
-	// Editable if the user is the scholar being viewed.
-	let uid = $derived(auth.getUserID());
-	let editable = $derived(currency !== null && uid !== null && currency.minters.includes(uid));
+	let newTokenOwner = $state<string | undefined>(undefined);
+	let newTokenCount = $state(1);
+	let newTokenConsent = $state(false);
+	let newTokenCreating = $state(false);
 </script>
 
 <Page title={currency ? currency.name : 'Oops'} subtitle="Currency">
 	{#if currency === null}
 		<Feedback error>Unknown currency.</Feedback>
 	{:else}
-		{#if editable}
+		{#if isMinter}
 			<EditableText
 				inline={false}
 				text={currency.description}
-				label="status"
 				placeholder="Explain the currency to others."
 				edit={(text) => db.updateCurrencyDescription(currency.id, text)}
 				note="Currency descriptions are public."
 			/>
 		{:else}
 			<p>{currency.description}</p>
-		{/if}
-		{#if count !== null}
-			<p>
-				There are <Tokens amount={count}></Tokens> tokens minted in this currency. <Link
-					to="/currency/{currency.id}/transactions">See all transactions</Link
-				>.
-			</p>
 		{/if}
 		<Cards>
 			<Card header="Minters">
@@ -92,8 +86,9 @@
 							tip="Add minter"
 							active={validEmail(newMinter) || ORCIDRegex.test(newMinter)}
 							action={async () => {
-								if (await handle(db.addCurrencyMinter(currency.id, currency.minters, newMinter)))
+								if (await handle(db.addCurrencyMinter(currency.id, currency.minters, newMinter))) {
 									newMinter = '';
+								}
 							}}>Add minter</Button
 						>
 					</form>
@@ -101,6 +96,62 @@
 						Minters can see, approve, and cancel transactions, and most importantly, mint new tokens
 						in this currency. They can also propose and improve currency exchanges and mergers.
 					</Note>
+				{/if}
+			</Card>
+			<Card header="Tokens">
+				<p>
+					There are {#if count !== null}<Tokens amount={count}></Tokens>{:else}an unknown number of{/if}
+					tokens minted in this currency, owned by <strong>{scholarCount}</strong> scholars and
+					<strong>{venueCount}</strong> distinct venues.
+				</p>
+				<p>
+					<Link to="/currency/{currency.id}/transactions">See all transactions</Link> in this currency.
+				</p>
+				{#if isMinter && venues}
+					<h3>Mint new tokens</h3>
+					<Feedback
+						>Be very careful when creating tokens. Too many tokens per scholar will diminish the
+						incentive to earn them.</Feedback
+					>
+					<form>
+						<label>
+							Which venue should own the new tokens?
+							<select bind:value={newTokenOwner}>
+								{#each venues as venue}<option value={venue.id}>{venue.title}</option
+									>{/each}</select
+							>
+						</label>
+						<Slider
+							min={1}
+							max={20}
+							value={newTokenCount}
+							step={1}
+							label="Number of new tokens"
+							change={(val) => (newTokenCount = val)}>{newTokenCount}</Slider
+						>
+						<Checkbox bind:on={newTokenConsent}>
+							I understand that creating excess tokens will erode this currencies value.</Checkbox
+						>
+
+						<Button
+							tip="Mint tokens"
+							active={!newTokenCreating &&
+								newTokenConsent &&
+								newTokenCount > 0 &&
+								newTokenOwner !== undefined}
+							action={async () => {
+								newTokenCreating = true;
+								if (
+									newTokenOwner !== undefined &&
+									(await handle(db.mintTokens(currency.id, newTokenCount, newTokenOwner)))
+								) {
+									newTokenCount = 0;
+									newTokenCreating = false;
+									newTokenConsent = false;
+								}
+							}}>Mint tokens</Button
+						>
+					</form>
 				{/if}
 			</Card>
 			<Card header="Venues">
@@ -115,9 +166,9 @@
 					<Feedback error>Unable to load venues.</Feedback>
 				{/if}
 			</Card>
-			{#if editable}
+			{#if isMinter}
 				<Card header="Settings" group="minters">
-					{#if editable}<EditableText
+					{#if isMinter}<EditableText
 							text={currency.name}
 							label="name"
 							placeholder="name"